+++ contracts/factory/src/error.rs
@@ -16,4 +16,5 @@ pub enum ContractError {
     SameAdmin = 109,
     NoAdminChangeInPlace = 110,
     AdminChangeExpired = 111,
+    TokenDecimalsInvalid = 112,
 }